requirements.txt
defusedxml==0.7.1
fonttools==4.38.0
fpdf2==2.6.1
mysql-connector-python==8.0.32
Pillow==9.4.0
protobuf==3.20.3
PyYAML==3.13
reportlab==3.6.12
